Kernel, Kernel Builder and other utilities.
The debian package for the kernel now extract and modify boot.scr instead
of creating a Symlink (Packages in the repositories are still using Symlink
until i finish to port all the build scripts to armStrap.

Kernel builder is back, armStrap doesn't use it for now but it can be call
manually with the -K switch to build all the debian packages for a Kernel.

U-Boot builder is somewhat back, it create the binaries but armStrap does
nothing with it yet. (-U switch)

RootFS Updater is now part of armStrap, it will allow more customisation of
the base rootfs in the futur.

I have learn how to write better commit message, but still need to learn
english :)