Add -I command line switch.
You can now add -I to -K when compiling a kernel, this will invoke
menuconfig to allow customisation of the kernel before building. If
you want, it will also save your new configuration as "custom". That
way you can rebuild the kernel with this new config another time or
use it as a starting point for more modifications.