Add -A cmdline switch to build every board/configuration
armStrap can now compile every kernels rootFS and U-Boot with just
one switch. Once done, it will call an hook script that can be used
to upload debian packages to a repository, or do wathever you want
with the result files.

The external script will be passed two parameters, first is where
all the packages are stores, second is the logfile.