First version that can actually build and boot a RaspBerry PI.
Various chances to make things more modular, better handling of
the fstab. Some cosmetic changes here and there. Disable the kernel
builder for the RaspBerry PI since it already have a very good, stable
and well maintained kernel. There is no point in re-inventing the wheel.

RaspBerry PI will stick with Raspbian as its rootFS since it's the only
one compiled for the ARMv6, debian and ubuntu target ARMv7.

The PI is the last piece of hardware i have to test, so don't expect new
board definition from me (I will probably buy a CubieTruck once it's out
and will add support for it). But i welcome any board definition if you are
willing to support it. Or, if you give me a board, i will write the board
definition for it :)