Always install the highest kernel version if no kernel version is selected.
Basically, what it does is a reverse sort of all the kernels avalable in the
repository and select the first that match the search pattern
${BOARD_KERNEL}-linux-${BOARD_KERNEL_CONFIG}-image-${BOARD_KERNEL_VERSION}
So if both sun4i-linux-default-image-3.4.75 and sun4i-linux-default-image-3.4.75+
are avalable, 3.4.75+ will be selected. By not specifying a kernel version in
config.sh, it will always match the highest kernel version avalable.